extract shared notebook data hook and dialog shell, split notes dialogs
